reset project in javascript
