+++ client/src/App.jsx
@@ -1,0 +1,18 @@
+
+import './App.css'
+import Navbar from './commons/Navbar/Navbar'
+import Footer from './commons/Footer/Footer'
+import HomePage from './pages/home/Home'
+
+function App() {
+
+  return (
+    <>
+      <Navbar></Navbar>
+      <HomePage></HomePage>
+      <Footer></Footer>
+    </>
+  )
+}
+
+export default App
