+++ client/src/commons/footer/Footer.jsx
@@ -1,0 +1,9 @@
+const Footer = ()=>{
+    return (<>
+        <footer className="py-5 bg-dark">
+            <div className="container"><p className="m-0 text-center text-white">Copyright &copy; Your Website 2022</p></div>
+        </footer>
+    </>)
+}
+
+export default Footer;
