--- client/src/components/footer/footer.component.tsx
@@ -1,10 +1,0 @@
-export const Footer = () => {
-    return (
-        <>
-            <footer  className="py-5 bg-dark">
-                <div  className="container"><p  className="m-0 text-center text-white">Copyright &copy; Your Website 2022</p></div>
-            </footer>
-            
-        </>
-    );
-}
