--- client/src/components/pages/Details/detail.page.tsx
@@ -1,6 +1,0 @@
-export const Details = ()=>{
-    return (
-        <>
-        </>
-    );
-}
