--- client/src/components/pages/home/home.page.css
@@ -1,4 +1,0 @@
-.inner-card{
-    top: 0.5rem; 
-    right: 0.5rem
-}
