+++ client/src/pages/home/Home.jsx
@@ -1,18 +1,6 @@
-import './home.page.css'
-
-export const Home = () => {
-    return (
-        <>
-            <header className="bg-dark py-5">
-                <div className="container px-4 px-lg-5 my-5">
-                    <div className="text-center text-white">
-                        <h1 className="display-4 fw-bolder">Shop in style</h1>
-                        <p className="lead fw-normal text-white-50 mb-0">With this shop hompeage template</p>
-                    </div>
-                </div>
-            </header>
-
-                   <section className="py-5">
+const HomePage = ()=>{
+    return (<>
+        <section className="py-5">
             <div className="container px-4 px-lg-5 mt-5">
                 <div className="row gx-4 gx-lg-5 row-cols-2 row-cols-md-3 row-cols-xl-4 justify-content-center">
                     <div className="col mb-5">
@@ -24,19 +12,20 @@ export const Home = () => {
                                 <div className="text-center">
                                     
                                     <h5 className="fw-bolder">Fancy Product</h5>
-                                  
+                                    
                                     $40.00 - $80.00
                                 </div>
                             </div>
+                            
                             <div className="card-footer p-4 pt-0 border-top-0 bg-transparent">
-                                <div className="text-center"><a className="btn btn-outline-dark mt-auto" href="#">View options</a></div>
+                                <div className="text-center"><a className="btn btn-outline-dark mt-auto" href="detail.html">View options</a></div>
                             </div>
                         </div>
                     </div>
                     <div className="col mb-5">
                         <div className="card h-100">
                             
-                            <div className="badge inner-card bg-dark text-white position-absolute"  >Sale</div>
+                            <div className="badge bg-dark text-white position-absolute inner-card" >Sale</div>
                             
                             <img className="card-img-top" src="https://dummyimage.com/450x300/dee2e6/6c757d.jpg" alt="..." />
                             
@@ -56,19 +45,20 @@ export const Home = () => {
                                     <span className="text-muted text-decoration-line-through">$20.00</span>
                                     $18.00
                                 </div>
-                            </div>pt
+                            </div>
+                            
                             <div className="card-footer p-4 pt-0 border-top-0 bg-transparent">
-                                <div className="text-center"><a className="btn btn-outline-dark mt-auto" href="#">Add to cart</a></div>
+                                <div className="text-center"><a className="btn btn-outline-dark mt-auto" href="detail.html">Add to cart</a></div>
                             </div>
                         </div>
                     </div>
                     <div className="col mb-5">
                         <div className="card h-100">
                             
-                            <div className="badge inner-card bg-dark text-white position-absolute"  >Sale</div>
+                            <div className="badge bg-dark text-white position-absolute inner-card" >Sale</div>
                             
                             <img className="card-img-top" src="https://dummyimage.com/450x300/dee2e6/6c757d.jpg" alt="..." />
-                           
+                            
                             <div className="card-body p-4">
                                 <div className="text-center">
                                     
@@ -80,7 +70,7 @@ export const Home = () => {
                             </div>
                             
                             <div className="card-footer p-4 pt-0 border-top-0 bg-transparent">
-                                <div className="text-center"><a className="btn btn-outline-dark mt-auto" href="#">Add to cart</a></div>
+                                <div className="text-center"><a className="btn btn-outline-dark mt-auto" href="detail.html">Add to cart</a></div>
                             </div>
                         </div>
                     </div>
@@ -101,6 +91,7 @@ export const Home = () => {
                                         <div className="bi-star-fill"></div>
                                         <div className="bi-star-fill"></div>
                                     </div>
+                                    
                                     $40.00
                                 </div>
                             </div>
@@ -113,7 +104,7 @@ export const Home = () => {
                     <div className="col mb-5">
                         <div className="card h-100">
                             
-                            <div className="badge inner-card bg-dark text-white position-absolute"  >Sale</div>
+                            <div className="badge bg-dark text-white position-absolute inner-card" >Sale</div>
                             
                             <img className="card-img-top" src="https://dummyimage.com/450x300/dee2e6/6c757d.jpg" alt="..." />
                             
@@ -139,12 +130,13 @@ export const Home = () => {
                             
                             <div className="card-body p-4">
                                 <div className="text-center">
-                                  
+                                    
                                     <h5 className="fw-bolder">Fancy Product</h5>
                                     
                                     $120.00 - $280.00
                                 </div>
                             </div>
+                            
                             <div className="card-footer p-4 pt-0 border-top-0 bg-transparent">
                                 <div className="text-center"><a className="btn btn-outline-dark mt-auto" href="#">View options</a></div>
                             </div>
@@ -153,10 +145,10 @@ export const Home = () => {
                     <div className="col mb-5">
                         <div className="card h-100">
                             
-                            <div className="badge inner-card bg-dark text-white position-absolute"  >Sale</div>
-                           
+                            <div className="badge bg-dark text-white position-absolute inner-card" >Sale</div>
+                            
                             <img className="card-img-top" src="https://dummyimage.com/450x300/dee2e6/6c757d.jpg" alt="..." />
-                           
+                            
                             <div className="card-body p-4">
                                 <div className="text-center">
                                     
@@ -169,11 +161,12 @@ export const Home = () => {
                                         <div className="bi-star-fill"></div>
                                         <div className="bi-star-fill"></div>
                                     </div>
-
+                                    
                                     <span className="text-muted text-decoration-line-through">$20.00</span>
                                     $18.00
                                 </div>
                             </div>
+                            
                             <div className="card-footer p-4 pt-0 border-top-0 bg-transparent">
                                 <div className="text-center"><a className="btn btn-outline-dark mt-auto" href="#">Add to cart</a></div>
                             </div>
@@ -181,7 +174,7 @@ export const Home = () => {
                     </div>
                     <div className="col mb-5">
                         <div className="card h-100">
-                           
+                            
                             <img className="card-img-top" src="https://dummyimage.com/450x300/dee2e6/6c757d.jpg" alt="..." />
                             
                             <div className="card-body p-4">
@@ -196,11 +189,11 @@ export const Home = () => {
                                         <div className="bi-star-fill"></div>
                                         <div className="bi-star-fill"></div>
                                     </div>
-
+                                    
                                     $40.00
                                 </div>
                             </div>
-
+                            
                             <div className="card-footer p-4 pt-0 border-top-0 bg-transparent">
                                 <div className="text-center"><a className="btn btn-outline-dark mt-auto" href="#">Add to cart</a></div>
                             </div>
@@ -208,7 +201,8 @@ export const Home = () => {
                     </div>
                 </div>
             </div>
-        </section> 
-        </>
-    )
+        </section>
+    </>)
 }
+
+export default HomePage;
