+++ client/src/pages/home/home.css
@@ -1,0 +1,3 @@
+.inner.card{
+    top: 0.5rem; right: 0.5rem
+}
