--- client/src/redux/slices/index.ts
@@ -1,1 +1,0 @@
-export {titlePage} from './title.slice';
